labs/02-langchain/demo-02-basic-llm-invocation-with-retry/pyproject.toml
[project]
name = "demo-02-basic-llm-invocation-with-retry"
version = "0.1.0"
description = "Basic LLM Invocation with Retry using LangChain and Gemini"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.128.0",
    "uvicorn==0.40.0",
    "python-dotenv==1.2.1",
    "pydantic==2.12.5",
    "langchain==1.2.8",
    "langchain-openai==1.1.7",
]
